requirements.txt
psycopg2-binary==2.9.11
python-dotenv==1.2.1
neo4j==6.0.2
altair==5.5.0
altex==0.2.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.55
botocore==1.40.55
cachetools==6.2.1
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.0
cycler==0.12.1
entrypoints==0.4
Faker==37.11.0
favicon==0.7.0
filelock==3.20.0
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
htbuilder==0.9.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.9
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.9.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
prometheus_client==0.23.1
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.8
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.50.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
